Extract Registerable module and memoize registered instances
