+++ lib/spandx/dotnet/parsers/sln.rb
@@ -4,7 +4,7 @@ module Spandx
   module Dotnet
     module Parsers
       class Sln < ::Spandx::Core::Parser
-        def self.matches?(filename)
+        def matches?(filename)
           filename.match?(/.*\.sln/)
         end
