+++ spec/unit/dotnet/parsers/csproj_spec.rb
@@ -34,8 +34,6 @@ RSpec.describe Spandx::Dotnet::Parsers::Csproj do
   end
 
   describe '.matches?' do
-    subject { described_class }
-
     specify { expect(subject.matches?('/root/simple.csproj')).to be(true) }
     specify { expect(subject.matches?('C:\Documents and Settings\simple.csproj')).to be(true) }
     specify { expect(subject.matches?('C:\Documents and Settings\hello world.csproj')).to be(true) }
